Week 3 Thursday Commit
Re-organized the repository so that it's easier to find things.

Chapter Reports are now in a single directory called "Chapter Reports", and auto-generated folders for those reports are now in sub-directories within "Chapter Reports".
(This folder is specifically distinguised 'chapter' because I feel it is very likely I will do other types of reports in the future).

Timecards are now in a single directory called "Timecards".

Images and *.csv* files that my reports (of any kind) need to pull from are now in a data_sources folder.

README is now updated and *.md* files will be hyperlinked once they are pushed to github.

Removed submodule.